+++ src/main/java/ellpeck/actuallyadditions/crafting/BlockCrafting.java
@@ -39,6 +39,7 @@ public class BlockCrafting{
     public static IRecipe recipeLiquidPlacer;
     public static IRecipe recipeLiquidCollector;
     public static IRecipe recipeCase;
+    public static IRecipe recipeIronCase;
     public static IRecipe recipeStoneCase;
     public static IRecipe recipeEnderCase;
     public static IRecipe recipeEnderPearlBlock;
@@ -83,8 +84,9 @@ public class BlockCrafting{
 
         //Laser Relay
         if(ConfigCrafting.LASER_RELAY.isEnabled()){
-            GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockLaserRelay),
-                    "ORO", "RCR", "ORO",
+            GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockLaserRelay, 2),
+                    "OBO", "RCR", "OBO",
+                    'B', new ItemStack(Blocks.redstone_block),
                     'O', new ItemStack(Blocks.obsidian),
                     'R', new ItemStack(Items.redstone),
                     'C', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal())));
@@ -155,6 +157,16 @@ public class BlockCrafting{
             recipeCase = Util.GetRecipes.lastIRecipe();
         }
 
+        //Iron Casing
+        if(ConfigCrafting.IRON_CASING.isEnabled()){
+            GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
+                    "WSW", "SQS", "WSW",
+                    'Q', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.QUARTZ.ordinal()),
+                    'W', "ingotIron",
+                    'S', "stickWood"));
+            recipeIronCase = Util.GetRecipes.lastIRecipe();
+        }
+
         //Ender Casing
         if(ConfigCrafting.ENDER_CASING.isEnabled()){
             GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.ENDER_CASING.ordinal()),
@@ -191,7 +203,7 @@ public class BlockCrafting{
                     "I I", "CAC", "I I",
                     'I', "ingotIron",
                     'C', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal()),
-                    'A', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal())));
+                    'A', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal())));
             recipeEnergizer = Util.GetRecipes.lastIRecipe();
         }
 
@@ -201,7 +213,7 @@ public class BlockCrafting{
                     " I ", "CAC", " I ",
                     'I', "ingotIron",
                     'C', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal()),
-                    'A', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal())));
+                    'A', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal())));
             recipeEnervator = Util.GetRecipes.lastIRecipe();
         }
 
@@ -209,7 +221,7 @@ public class BlockCrafting{
         if(ConfigCrafting.LAVA_FACTORY.isEnabled()){
             GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockLavaFactoryController),
                     "SCS", "ISI", "LLL",
-                    'C', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal()),
+                    'C', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
                     'S', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal()),
                     'I', "blockIron",
                     'L', Items.lava_bucket));
@@ -220,7 +232,7 @@ public class BlockCrafting{
         if(ConfigCrafting.CASING.isEnabled()){
             GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockMisc, 32, TheMiscBlocks.LAVA_FACTORY_CASE.ordinal()),
                     "ICI",
-                    'C', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal()),
+                    'C', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
                     'I', "blockIron"));
             recipeCasing = Util.GetRecipes.lastIRecipe();
         }
@@ -241,7 +253,7 @@ public class BlockCrafting{
             GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockOreMagnet),
                     "SSS", "CBC", "ISI",
                     'S', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.LAVA_FACTORY_CASE.ordinal()),
-                    'C', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal()),
+                    'C', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
                     'B', new ItemStack(InitItems.itemBatteryDouble),
                     'I', new ItemStack(Blocks.iron_block)));
             recipeMiner = Util.GetRecipes.lastIRecipe();
@@ -347,7 +359,7 @@ public class BlockCrafting{
                     "IEI", "GLG", "ICI",
                     'I', "ingotIron",
                     'G', "ingotGold",
-                    'E', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal()),
+                    'E', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
                     'L', "treeLeaves",
                     'C', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal())));
             recipeLeafGen = Util.GetRecipes.lastIRecipe();
@@ -403,7 +415,7 @@ public class BlockCrafting{
                     "IQI", "CDC", "IBI",
                     'D', "blockDiamond",
                     'I', "ingotIron",
-                    'Q', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal()),
+                    'Q', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
                     'C', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal()),
                     'B', new ItemStack(Blocks.iron_bars)));
             recipeSolar = Util.GetRecipes.lastIRecipe();
@@ -413,9 +425,9 @@ public class BlockCrafting{
         if(ConfigCrafting.HEAT_COLLECTOR.isEnabled()){
             GameRegistry.addRecipe(new ShapedOreRecipe(new ItemStack(InitBlocks.blockHeatCollector),
                     "BRB", "CDC", "BQB",
-                    'D', "gemDiamond",
+                    'D', "ingotIron",
                     'R', new ItemStack(Items.repeater),
-                    'Q', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.STONE_CASING.ordinal()),
+                    'Q', new ItemStack(InitBlocks.blockMisc, 1, TheMiscBlocks.IRON_CASING.ordinal()),
                     'L', new ItemStack(Items.lava_bucket),
                     'C', new ItemStack(InitItems.itemMisc, 1, TheMiscItems.COIL_ADVANCED.ordinal()),
                     'B', new ItemStack(Blocks.iron_bars)));