+++ src/main/java/ellpeck/actuallyadditions/items/metalists/TheMiscItems.java
@@ -29,7 +29,8 @@ public enum TheMiscItems{
     RICE_SLIME("RiceSlime", EnumRarity.uncommon),
     CANOLA("Canola", EnumRarity.uncommon),
     CUP("Cup", EnumRarity.uncommon),
-    BAT_WING("BatWing", EnumRarity.rare);
+    BAT_WING("BatWing", EnumRarity.rare),
+    DRILL_CORE("DrillCore", EnumRarity.uncommon);
 
     public final String name;
     public final EnumRarity rarity;